# contracts/openapi.yaml
openapi: 3.1.0
info: { title: FIAP X Video API, version: 0.1.0 }
servers: [ { url: 'http://localhost:8080' } ]
paths:
...
  /v1/jobs:
    get:
      summary: Lista jobs do proprietário por cursor
      security: [ { bearerAuth: [ ] } ]
      parameters: [ { name: cursor, in: query, required: false, schema: { type: string } }, { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } } ]
      responses: { '200': { description: Página de jobs, content: { application/json: { schema: { $ref: '#/components/schemas/JobPage' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '401': { $ref: '#/components/responses/Unauthorized' } }
    post:
      summary: Cria um job a partir de um objeto já enviado
      security: [ { bearerAuth: [ ] } ]
      parameters: [ { name: Idempotency-Key, in: header, required: false, description: Chave opaca (até 128 caracteres) para repetir a criação com segurança, schema: { type: string, maxLength: 128 } } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [ sourceKey ], properties: { sourceKey: { type: string } } } } } }
      responses: { '201': { description: Job criado ou recuperado para a mesma chave de idempotência, content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '401': { $ref: '#/components/responses/Unauthorized' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' } }
  /v1/jobs/{id}:
    get:
      summary: Consulta o estado de um job
      security: [ { bearerAuth: [ ] } ]
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '200': { description: Estado atual, content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } } }, '401': { $ref: '#/components/responses/Unauthorized' } }
  /v1/jobs/{id}/download:
    get:
      summary: Obtém acesso temporário ao resultado do proprietário
      security: [ { bearerAuth: [ ] } ]
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '200': { description: URL temporária, content: { application/json: { schema: { $ref: '#/components/schemas/DownloadResult' } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' }, '503': { $ref: '#/components/responses/StorageUnavailable' } }
...
components:
  securitySchemes: { bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT } }
...
  responses:
    ValidationError: { description: Dados inválidos, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
    Unauthorized: { description: Bearer token inválido ou ausente, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
    NotFound: { description: Vídeo ou recurso não encontrado para o proprietário, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    Conflict: { description: Estado incompatível, e-mail já cadastrado, chave de idempotência reutilizada com outro payload ou vídeo não confirmado, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    StorageUnavailable: { description: Object storage temporariamente indisponível, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
  schemas:
...
    Problem: { type: object, properties: { type: { type: string }, title: { type: string }, status: { type: integer }, detail: { type: string }, instance: { type: string }, code: { type: string } } }
    Job: { type: object, properties: { id: { type: string, format: uuid }, userId: { type: string, format: uuid }, sourceKey: { type: string }, resultKey: { type: [ string, 'null' ] }, status: { enum: [ PENDING, PROCESSING, COMPLETED, FAILED ] }, createdAt: { type: string, format: date-time } } }
    JobPage: { type: object, required: [ items ], properties: { items: { type: array, items: { $ref: '#/components/schemas/Job' } }, nextCursor: { type: [ string, 'null' ] } } }
    DownloadResult: { type: object, required: [ downloadUrl, expiresAt, filename, contentType, sizeBytes ], properties: { downloadUrl: { type: string, format: uri }, expiresAt: { type: string, format: date-time }, filename: { type: string }, contentType: { type: string }, sizeBytes: { type: integer } } }